requirements.txt
streamlit>=1.36,<1.40
requests>=2.32
python-dotenv>=1.0
firebase-admin>=6.5

